+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 1.4.2 - Upcoming Release
+
+- Added
+  - Multi signal select
+  - Multi signal reorder
+- Fixed
+  Drag and drop from tree view places signals at the end of group
+
 ## 1.4.1 - 9/12/2025 - Latest Release
 
 - Added